# public/openapi.json
{"openapi":"3.0.0","info":{"title":"iDoc API Reference","version":"","description":"Documentacion de la Api","termsOfService":"","license":null,"contact":[],"x-logo":{"url":"","altText":"iDoc API Reference","backgroundColor":""}},"components":{"securitySchemes":{"BearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":[]},"servers":[{"url":"http:\/\/127.0.0.1:8000","description":"Documentation generator server."},{"url":"http:\/\/test.example.com","description":"Test server."},{"url":"http:\/\/127.0.0.1:8000","description":"Test Local server."}],"paths":{"\/api\/auth\/login":{"post":{"tags":["Autentificacion"],"operationId":"Login de un usuario","description":"Proceso para obtener la autorizacion para consumir las apis","requestBody":{"description":"Proceso para obtener la autorizacion para consumir las apis","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","example":"admin@gmail.com","description":"Email del Usuario"},"password":{"type":"string","example":"12345678","description":"Password del Usuario"}},"example":{"email":"admin@gmail.com","password":"12345678"}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"status":true,"message":"Login correcto.","errors":[],"data":{"token":"5|iPkrWUHEEJrAZD8u1MS2jvTfYqSRqRMpCm2xdUCz0380c43b","expiration_date":"2025-02-19T14:11:14.968924Z","user":[]},"others":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/127.0.0.1:8000\/api\/auth\/login\" \n    -d \"email\"=\"admin@gmail.com\" \\\n    -d \"password\"=\"12345678\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/127.0.0.1:8000\/api\/auth\/login\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"email\": \"admin@gmail.com\",\n    \"password\": \"12345678\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"email\" =>  \"admin@gmail.com\",\n    \"password\" =>  \"12345678\",\n];\n\n\n$ch = curl_init(\"http:\/\/127.0.0.1:8000\/api\/auth\/login\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/categorias":{"get":{"tags":["Categorias"],"operationId":"Listado de las Categorias","description":"Obtiene los datos de las categorias","parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"ststus":true,"message":"Datos obtenidos","errors":[],"data":[{"nombre":"Ropa","descripcion":"Descripcion de la Ropa","creado":"2024-08-16 14:05:16","actualizado":"2024-08-16 14:05:16"}],"others":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/127.0.0.1:8000\/api\/categorias\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/127.0.0.1:8000\/api\/categorias\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/127.0.0.1:8000\/api\/categorias\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"post":{"tags":["Categorias"],"operationId":"Crear una Categoria","description":"Endpoint para crear una categoria","requestBody":{"description":"Endpoint para crear una categoria","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["nombre"],"properties":{"nombre":{"type":"string","example":"Ropa","description":"Nombre de la categoria que deseamos crear"},"descripcion":{"type":"Descripcon","example":"Ropa para todo tipo de edad","description":"de la categoria"}},"example":{"nombre":"Ropa","descripcion":"Ropa para todo tipo de edad"}}}}},"parameters":[],"responses":{"201":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"ststus":true,"message":"Datos obtenidos","errors":[],"data":{"nombre":"Ropa","descripcion":"Descripcion de la Ropa","creado":"2024-08-16 14:05:16","actualizado":"2024-08-16 14:05:16"},"others":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/127.0.0.1:8000\/api\/categorias\" \n    -d \"nombre\"=\"Ropa\" \\\n    -d \"descripcion\"=\"Ropa para todo tipo de edad\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/127.0.0.1:8000\/api\/categorias\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"nombre\": \"Ropa\",\n    \"descripcion\": \"Ropa para todo tipo de edad\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"nombre\" =>  \"Ropa\",\n    \"descripcion\" =>  \"Ropa para todo tipo de edad\",\n];\n\n\n$ch = curl_init(\"http:\/\/127.0.0.1:8000\/api\/categorias\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/categorias\/{id}":{"get":{"tags":["Categorias"],"operationId":"Obtiene una Categoria","description":"Obtiene una categoria por ID","parameters":[{"in":"path","name":"id","description":"de la categoria que deseamos obtener","required":false,"schema":{"type":"Id","example":"1"}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"ststus":true,"message":"Datos obtenidos","errors":[],"data":{"nombre":"Ropa","descripcion":"Descripcion de la Ropa","creado":"2024-08-16 14:05:16","actualizado":"2024-08-16 14:05:16"},"others":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/127.0.0.1:8000\/api\/categorias\/{id}\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/127.0.0.1:8000\/api\/categorias\/{id}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/127.0.0.1:8000\/api\/categorias\/{id}\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"put":{"tags":["Categorias"],"operationId":"Actualizar una Categoria","description":"Endpoint para actualizar una categoria","requestBody":{"description":"Endpoint para actualizar una categoria","required":true,"content":{"application\/json":{"schema":{"type":"object"}}}},"parameters":[{"in":"path","name":"id","description":"de la categoria que deseamos actualizar","required":false,"schema":{"type":"Id","example":"1"}}],"responses":{"201":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"ststus":true,"message":"Datos obtenidos","errors":[],"data":{"nombre":"Ropa","descripcion":"Descripcion de la Ropa","creado":"2024-08-16 14:05:16","actualizado":"2024-08-16 14:05:16"},"others":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X PUT \"http:\/\/127.0.0.1:8000\/api\/categorias\/{id}\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/127.0.0.1:8000\/api\/categorias\/{id}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"PUT\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/127.0.0.1:8000\/api\/categorias\/{id}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"PUT\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"delete":{"tags":["Categorias"],"operationId":"Eliminar una Categoria","description":"Endpoint para elimiar una categoria","parameters":[{"in":"path","name":"id","description":"de la categoria que deseamos elimiar","required":false,"schema":{"type":"Id","example":"1"}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":[]}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X DELETE \"http:\/\/127.0.0.1:8000\/api\/categorias\/{id}\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/127.0.0.1:8000\/api\/categorias\/{id}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"DELETE\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/127.0.0.1:8000\/api\/categorias\/{id}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"DELETE\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/productos":{"get":{"tags":["Productos"],"operationId":"Listado de las Productos","description":"Obtiene los datos de las productos","parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"ststus":true,"message":"Datos obtenidos","errors":[],"data":[{"nombre":"Producto 5 Editado","descripcion":"Descripcion 1","precio":10,"cantidad":1,"image":"http:\/\/127.0.0.1:8000\/storage\/1723851231.png","category":{"nombre":"Electronica","descripcion":"Articulos electronicos para telefonos, computadoras, etc","creado":null,"actualizado":null}}],"others":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/127.0.0.1:8000\/api\/productos\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/127.0.0.1:8000\/api\/productos\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/127.0.0.1:8000\/api\/productos\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"post":{"tags":["Productos"],"operationId":"Crear un Producto","description":"Obtiene los datos de las productos","requestBody":{"description":"Obtiene los datos de las productos","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["nombre","descripcion","category_id","precio","cantidad","thumbnail"],"properties":{"nombre":{"type":"string","example":"Polera Edicion 2024","description":"Nombre del Producto"},"descripcion":{"type":"string","example":"Descripcion","description":"Descripcion del Producto"},"category_id":{"type":"integer","example":1,"description":"categoria_id del Producto"},"precio":{"type":"number","example":10.45,"description":"Precio del Producto"},"cantidad":{"type":"integer","example":10,"description":"Cantidad del Producto"},"thumbnail":{"type":"image","example":"Z1=E2mdNRd\"B","description":""}},"example":{"nombre":"Polera Edicion 2024","descripcion":"Descripcion","category_id":1,"precio":10.45,"cantidad":10,"thumbnail":"Z1=E2mdNRd\"B"}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"ststus":true,"message":"Datos obtenidos","errors":[],"data":{"nombre":"Producto 5 Editado","descripcion":"Descripcion 1","precio":10,"cantidad":1,"image":"http:\/\/127.0.0.1:8000\/storage\/1723851231.png","category":{"nombre":"Electronica","descripcion":"Articulos electronicos para telefonos, computadoras, etc","creado":null,"actualizado":null}},"others":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/127.0.0.1:8000\/api\/productos\" \n    -d \"nombre\"=\"Polera Edicion 2024\" \\\n    -d \"descripcion\"=\"Descripcion\" \\\n    -d \"category_id\"=\"1\" \\\n    -d \"precio\"=\"10.45\" \\\n    -d \"cantidad\"=\"10\" \\\n    -d \"thumbnail\"=\"Z1=E2mdNRd\"B\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/127.0.0.1:8000\/api\/productos\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"nombre\": \"Polera Edicion 2024\",\n    \"descripcion\": \"Descripcion\",\n    \"category_id\": \"1\",\n    \"precio\": \"10.45\",\n    \"cantidad\": \"10\",\n    \"thumbnail\": \"Z1=E2mdNRd&quot;B\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"nombre\" =>  \"Polera Edicion 2024\",\n    \"descripcion\" =>  \"Descripcion\",\n    \"category_id\" =>  \"1\",\n    \"precio\" =>  \"10.45\",\n    \"cantidad\" =>  \"10\",\n    \"thumbnail\" =>  \"Z1=E2mdNRd&quot;B\",\n];\n\n\n$ch = curl_init(\"http:\/\/127.0.0.1:8000\/api\/productos\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/productos\/{id}":{"get":{"tags":["Productos"],"operationId":"Obtener un producto","description":"Obtiene los datos de las productos","parameters":[{"in":"path","name":"id","description":"del producto que deseamos obtener","required":false,"schema":{"type":"Id","example":"1"}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"ststus":true,"message":"Datos obtenidos","errors":[],"data":[{"nombre":"Ropa","descripcion":"Descripcion de la Ropa","creado":"2024-08-16 14:05:16","actualizado":"2024-08-16 14:05:16"}],"others":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/127.0.0.1:8000\/api\/productos\/{id}\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/127.0.0.1:8000\/api\/productos\/{id}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/127.0.0.1:8000\/api\/productos\/{id}\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"put":{"tags":["Productos"],"operationId":"Actualizar una Producto","description":"Endpoint para actualizar una producto","requestBody":{"description":"Endpoint para actualizar una producto","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["nombre","descripcion","category_id","precio","cantidad","thumbnail"],"properties":{"nombre":{"type":"string","example":"Polera Edicion 2024","description":"Nombre del Producto"},"descripcion":{"type":"string","example":"Descripcion","description":"Descripcion del Producto"},"category_id":{"type":"integer","example":1,"description":"categoria_id del Producto"},"precio":{"type":"number","example":10.45,"description":"Precio del Producto"},"cantidad":{"type":"integer","example":10,"description":"Cantidad del Producto"},"thumbnail":{"type":"image","example":"`7f}[}nXL9;7","description":""}},"example":{"nombre":"Polera Edicion 2024","descripcion":"Descripcion","category_id":1,"precio":10.45,"cantidad":10,"thumbnail":"`7f}[}nXL9;7"}}}}},"parameters":[{"in":"path","name":"id","description":"de la producto que deseamos actualizar","required":false,"schema":{"type":"Id","example":"1"}}],"responses":{"201":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"ststus":true,"message":"Datos obtenidos","errors":[],"data":{"nombre":"Producto 5 Editado","descripcion":"Descripcion 1","precio":10,"cantidad":1,"image":"http:\/\/127.0.0.1:8000\/storage\/1723851231.png","category":{"nombre":"Electronica","descripcion":"Articulos electronicos para telefonos, computadoras, etc","creado":null,"actualizado":null}},"others":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X PUT \"http:\/\/127.0.0.1:8000\/api\/productos\/{id}\" \n    -d \"nombre\"=\"Polera Edicion 2024\" \\\n    -d \"descripcion\"=\"Descripcion\" \\\n    -d \"category_id\"=\"1\" \\\n    -d \"precio\"=\"10.45\" \\\n    -d \"cantidad\"=\"10\" \\\n    -d \"thumbnail\"=\"`7f}[}nXL9;7\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/127.0.0.1:8000\/api\/productos\/{id}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"nombre\": \"Polera Edicion 2024\",\n    \"descripcion\": \"Descripcion\",\n    \"category_id\": \"1\",\n    \"precio\": \"10.45\",\n    \"cantidad\": \"10\",\n    \"thumbnail\": \"`7f}[}nXL9;7\",\n})\n\nfetch(url, {\n    method: \"PUT\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"nombre\" =>  \"Polera Edicion 2024\",\n    \"descripcion\" =>  \"Descripcion\",\n    \"category_id\" =>  \"1\",\n    \"precio\" =>  \"10.45\",\n    \"cantidad\" =>  \"10\",\n    \"thumbnail\" =>  \"`7f}[}nXL9;7\",\n];\n\n\n$ch = curl_init(\"http:\/\/127.0.0.1:8000\/api\/productos\/{id}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"PUT\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"delete":{"tags":["Productos"],"operationId":"Eliminar un Producto","description":"Endpoint para elimiar una producto","parameters":[{"in":"path","name":"id","description":"de la producto que deseamos elimiar","required":false,"schema":{"type":"Id","example":"1"}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":[]}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X DELETE \"http:\/\/127.0.0.1:8000\/api\/productos\/{id}\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/127.0.0.1:8000\/api\/productos\/{id}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"DELETE\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/127.0.0.1:8000\/api\/productos\/{id}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"DELETE\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}}},"x-tagGroups":[]}